Remove unnecessary page loader functions
