+++ src/_app/index.ts
@@ -1,3 +1,3 @@
 export { default as fetchAppData } from './fetchAppData';
 export { default as renderApp, renderEnhancedApp } from './render';
-export { getAppFile, getSingleAppFile } from './getAppFile';
+export { getAppFile } from './getAppFile';
