+++ src/_document/index.ts
@@ -1,2 +1,2 @@
 export { default as serverRenderDocument } from './render';
-export { getSingleDocumentFile } from './getDocumentFile';
+export { getDocumentFile } from './getDocumentFile';
