+++ src/_error/index.ts
@@ -1,0 +1,2 @@
+export { getErrorPagePath } from './getErrorPagePath';
+export { InternalError } from './InternalError';
