+++ src/loadFile.ts
@@ -1,27 +1,7 @@
-import { requireAsIfOnServer } from './server';
-import type { PageFile } from './commonTypes';
-
-export function loadSingleFile<FileType>({
+export function loadFile<FileType>({
   absolutePath,
 }: {
   absolutePath: string;
 }): FileType {
   return require(absolutePath);
-}
-
-export function loadFile<FileType>({
-  absolutePath,
-  nonIsolatedModules,
-}: {
-  absolutePath: string;
-  nonIsolatedModules: string[];
-}): PageFile<FileType> {
-  return {
-    client: require(absolutePath),
-    server: requireAsIfOnServer<FileType>({
-      path: absolutePath,
-      nonIsolatedModules,
-    }),
-    path: absolutePath,
-  };
 }
