+++ src/page/index.ts
@@ -1,8 +1,6 @@
 export { getPageInfo } from './getPageInfo';
 export { makePageElement } from './makePageElement';
 export {
-  getSinglePageFile,
-  getSinglePageFileIfExists,
   getPageFile,
   getPageFileIfExists,
   getPagePath,
